backend/requirements.txt
requests
alembic
SQLAlchemy
PyJWT
uvicorn
fastapi
gunicorn
fastapi-event
pythondi
click
greenlet
python-multipart
bcrypt
pydantic
aiomysql
aiosqlite
ujson
python-dotenv